test coordinate bad vibes

+++ src/main/java/it/unibo/model/api/SceneBuilder.java
@@ -32,5 +32,5 @@ public interface SceneBuilder {
     /**
      * @return the offset for the x axis
      */
-    double offsetX();
+    int offsetX();
 }
